// types/IFixedIncomeNote.ts
export namespace SubscriptionSettledEvent {
  export type InputTuple = [
    orderId: BytesLike,
    investor: AddressLike,
    amount: BigNumberish,
    currency: string,
    settlementRef: string,
    settlementNetwork: string
  ];
  export type OutputTuple = [
    orderId: string,
    investor: string,
    amount: bigint,
    currency: string,
    settlementRef: string,
    settlementNetwork: string
  ];
  export interface OutputObject {
    orderId: string;
    investor: string;
    amount: bigint;
    currency: string;
    settlementRef: string;
    settlementNetwork: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace CouponPaidEvent {
  export type InputTuple = [
    periodId: BigNumberish,
    grossAmount: BigNumberish,
    withholding: BigNumberish,
    netAmount: BigNumberish,
    settlementRef: string,
    settlementNetwork: string
  ];
  export type OutputTuple = [
    periodId: bigint,
    grossAmount: bigint,
    withholding: bigint,
    netAmount: bigint,
    settlementRef: string,
    settlementNetwork: string
  ];
  export interface OutputObject {
    periodId: bigint;
    grossAmount: bigint;
    withholding: bigint;
    netAmount: bigint;
    settlementRef: string;
    settlementNetwork: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace RedemptionPaidEvent {
  export type InputTuple = [
    amount: BigNumberish,
    settlementRef: string,
    settlementNetwork: string
  ];
  export type OutputTuple = [
    amount: bigint,
    settlementRef: string,
    settlementNetwork: string
  ];
  export interface OutputObject {
    amount: bigint;
    settlementRef: string;
    settlementNetwork: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}